requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCL==1.3.0
discord.py[voice]
dnspython==1.16.0
youtube_dl
asyncio
praw
basc_py4chan
pymongo
DiscordUtils
googletrans==3.1.0a0
pafy
ffmpeg
numpy==1.19.3
matplotlib